fix(pkg/scripts_test): darwin tests
